fix(rpc): getblock verbosity 2 side-chain panic (GHSA-x6v8-c2xp-928m) (#10889)
* test(rpc): add regression test for getblock side-chain panic (GHSA-x6v8-c2xp-928m)

* fix(rpc): use i64 for transaction confirmations to avoid side-chain panic (GHSA-x6v8-c2xp-928m)

* fix(rpc): resolve clippy and rustfmt warnings

Remove unused `mut` on mempool mock and apply rustfmt formatting.

---------

(cherry picked from commit b5e122f5f9c0aa490a3d5f2e8e4000d2a6023d73)

+++ zebra-rpc/src/methods/types/transaction.rs
@@ -298,7 +298,7 @@ pub struct TransactionObject {
     /// mempool.
     #[serde(skip_serializing_if = "Option::is_none")]
     #[getter(copy)]
-    pub(crate) confirmations: Option<u32>,
+    pub(crate) confirmations: Option<i64>,
 
     /// Transparent inputs of the transaction.
     #[serde(rename = "vin")]
@@ -890,7 +890,7 @@ impl TransactionObject {
     pub fn from_transaction(
         tx: Arc<Transaction>,
         height: Option<block::Height>,
-        confirmations: Option<u32>,
+        confirmations: Option<i64>,
         network: &Network,
         block_time: Option<DateTime<Utc>>,
         block_hash: Option<block::Hash>,
